Change how centered menu width is calculated
min_width is changed from being an arbitrary compile-time value to being
defined as the prompt width.

The prompt width is no longer added to the width of the longest line to
determine the menu width. This had resulted in lines being a lot longer
than they needed to be in some cases. A downside of this is that the
input window is now sometimes smaller than can be read.